feat(manifest): add schema definition
