+++ src/hooks/useAuth.ts
@@ -1,7 +1,7 @@
 'use client';
 
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { useAuthStore } from '@/store/authStore';
+import { useAuthStore } from '../store/authStore';
 import { useRouter } from 'next/navigation';
 
 interface Credentials {
