+++ src/lib/moduleManifest.ts
@@ -1,60 +1,9 @@
-import Ajv, { JSONSchemaType } from 'ajv';
+import Ajv from 'ajv';
 import SchemaDefinition from '../services/schemaRegistry/schemaDefinition.model';
-
-export interface ModuleManifest {
-  name: string;
-  version: string;
-  description?: string;
-  endpoints: {
-    rest: string;
-    ws?: string;
-  };
-  schemas: Record<string, any>;
-  dependencies?: string[];
-  events?: {
-    publish?: string[];
-    subscribe?: string[];
-  };
-}
-
-// JSON Schema definition describing the shape of a module manifest.
-// The schema is used by AJV to validate manifests at runtime and lists
-// all required fields along with optional ones like `dependencies` or
-// `events`.
-const manifestSchema: JSONSchemaType<ModuleManifest> = {
-  type: 'object',
-  properties: {
-    name: { type: 'string' },
-    version: { type: 'string' },
-    description: { type: 'string', nullable: true },
-    endpoints: {
-      type: 'object',
-      properties: {
-        rest: { type: 'string' },
-        ws: { type: 'string', nullable: true },
-      },
-      required: ['rest'],
-      additionalProperties: false,
-    },
-    schemas: { type: 'object', additionalProperties: true },
-    dependencies: {
-      type: 'array',
-      items: { type: 'string' },
-      nullable: true,
-    },
-    events: {
-      type: 'object',
-      properties: {
-        publish: { type: 'array', items: { type: 'string' }, nullable: true },
-        subscribe: { type: 'array', items: { type: 'string' }, nullable: true },
-      },
-      additionalProperties: false,
-      nullable: true,
-    },
-  },
-  required: ['name', 'version', 'endpoints', 'schemas'],
-  additionalProperties: false,
-};
+import {
+  moduleManifestSchema,
+  type ModuleManifest,
+} from './moduleManifestSchema';
 
 // Configure AJV to collect all validation errors instead of stopping after the
 // first one. This produces more actionable feedback for developers.
@@ -87,7 +36,7 @@ const ajv = new Ajv({ allErrors: true });
  * message. Presenting distilled messages instead of raw AJV objects helps keep
  * validation feedback actionable.
  */
-export const validateManifest = ajv.compile(manifestSchema);
+export const validateManifest = ajv.compile(moduleManifestSchema);
 
 export async function approveManifest(manifest: ModuleManifest): Promise<boolean> {
   const valid = validateManifest(manifest);
